+++ src/plans/gradient_plan.jl
@@ -336,35 +336,6 @@ IdentityUpdateRule(M) = IdentityUpdateRule() # Temp hack for the mockup
 Gradient() = DirectionUpdateRuleFactory(IdentityUpdateRule)
 
 """
-    MomentumGradient()
-
-Append a momentum to a gradient processor, where the last direction and last iterate are
-stored and the new is composed as ``η_i = m*η_{i-1}' - s d_i``,
-where ``sd_i`` is the current (inner) direction and ``η_{i-1}'`` is the vector transported
-last direction multiplied by momentum ``m``.
-
-# Input
-
-* `M` (optional)
-
-# Keyword arguments
-
-* $(_kw_p_default)
-* `direction_update=`[`IdentityUpdateRule`](@ref) preprocess the actual gradient before adding momentum
-* $(_kw_X_default)`
-* `momentum=0.2` amount of momentum to use
-* `vector_transport_method=default_vector_transport_method(M, typeof(p)),
-
-!!! info
-    This function generates a [`DirectionUpdateRuleFactory`](@ref) for this
-    for the [`MomentumGradientRule`](@ref). If you do not provide the manifold
-    or any other of the keywords that require the manifold for a default,
-    you can provide a manifold later to use that manifold's default's instead.
-"""
-MomentumGradient(args...; kwargs...) =
-    DirectionUpdateRuleFactory(MomentumGradientRule, args...; kwargs...)
-
-"""
     MomentumGradientRule <: DirectionUpdateRule
 
 
@@ -404,14 +375,14 @@ mutable struct MomentumGradientRule{P,T,R<:Real,VTM<:AbstractVectorTransportMeth
     X_old::T
 end
 function MomentumGradientRule(
-    M::AbstractManifold,
-    p::P=rand(M);
+    M::AbstractManifold;
+    p::P=rand(M),
     direction::DirectionUpdateRule=IdentityUpdateRule(),
     vector_transport_method::VTM=default_vector_transport_method(M, typeof(p)),
-    X=zero_vector(M, p),
-    momentum=0.2,
-) where {P,VTM<:AbstractVectorTransportMethod}
-    return MomentumGradientRule{P,typeof(X),typeof(momentum),VTM}(
+    X::Q=zero_vector(M, p),
+    momentum::F=0.2,
+) where {P,Q,F<:Real,VTM<:AbstractVectorTransportMethod}
+    return MomentumGradientRule{P,Q,F,VTM}(
         momentum, p, direction, vector_transport_method, X
     )
 end
@@ -430,7 +401,36 @@ function (mg::MomentumGradientRule)(
 end
 
 """
-    AverageGradient <: DirectionUpdateRule
+    MomentumGradient()
+
+Append a momentum to a gradient processor, where the last direction and last iterate are
+stored and the new is composed as ``η_i = m*η_{i-1}' - s d_i``,
+where ``sd_i`` is the current (inner) direction and ``η_{i-1}'`` is the vector transported
+last direction multiplied by momentum ``m``.
+
+# Input
+
+* `M` (optional)
+
+# Keyword arguments
+
+* $(_kw_p_default)
+* `direction=`[`IdentityUpdateRule`](@ref) preprocess the actual gradient before adding momentum
+* $(_kw_X_default)`
+* `momentum=0.2` amount of momentum to use
+* $(_kw_vector_transport_method_default): $(_kw_vector_transport_method)
+
+!!! info
+    This function generates a [`DirectionUpdateRuleFactory`](@ref) for this
+    for the [`MomentumGradientRule`](@ref). If you do not provide the manifold
+    or any other of the keywords that require the manifold for a default,
+    you can provide a manifold later to use that manifold's default's instead.
+"""
+MomentumGradient(args...; kwargs...) =
+    DirectionUpdateRuleFactory(MomentumGradientRule, args...; kwargs...)
+
+"""
+    AverageGradientRule <: DirectionUpdateRule
 
 Add an average of gradients to a gradient processor. A set of previous directions (from the
 inner processor) and the last iterate are stored, average is taken after vector transporting
@@ -443,11 +443,12 @@ them to the current iterates tangent space.
 * `vector_transport_method`: vector transport method to use
 
 # Constructors
-    AverageGradient(
-        M::AbstractManifold,
+
+    AverageGradientRule(
+        M::AbstractManifold;
         p::P=rand(M);
         n::Int=10
-        s::DirectionUpdateRule=IdentityUpdateRule();
+        direction::DirectionUpdateRule=IdentityUpdateRule();
         gradients = fill(zero_vector(p.M, o.x),n),
         last_iterate = deepcopy(x0),
         vector_transport_method = default_vector_transport_method(M, typeof(p))
@@ -456,104 +457,114 @@ them to the current iterates tangent space.
 Add average to a gradient problem, where
 
 * `n`:                       determines the size of averaging
-* `s`:                       is the internal [`DirectionUpdateRule`](@ref) to determine the gradients to store
+* `direction`:                       is the internal [`DirectionUpdateRule`](@ref) to determine the gradients to store
 * `gradients`:               can be pre-filled with some history
 * `last_iterate`:            stores the last iterate
 * `vector_transport_method`: determines how to transport all gradients to the current iterates tangent space before averaging
 """
-mutable struct AverageGradient{P,T,VTM<:AbstractVectorTransportMethod} <:
+mutable struct AverageGradientRule{P,T,VTM<:AbstractVectorTransportMethod} <:
                DirectionUpdateRule
     gradients::AbstractVector{T}
     last_iterate::P
     direction::DirectionUpdateRule
     vector_transport_method::VTM
 end
-function AverageGradient(
-    M::AbstractManifold,
-    p::P=rand(M);
+function AverageGradientRule(
+    M::AbstractManifold;
+    p::P=rand(M),
     n::Int=10,
     direction::DirectionUpdateRule=IdentityUpdateRule(),
     gradients=[zero_vector(M, p) for _ in 1:n],
     vector_transport_method::VTM=default_vector_transport_method(M, typeof(p)),
 ) where {P,VTM}
-    return AverageGradient{P,eltype(gradients),VTM}(
-        gradients, p, direction, vector_transport_method
+    return AverageGradientRule{P,eltype(gradients),VTM}(
+        gradients, copy(M, p), direction, vector_transport_method
     )
 end
-function (a::AverageGradient)(mp::AbstractManoptProblem, s::AbstractGradientSolverState, k)
+function (a::AverageGradientRule)(
+    mp::AbstractManoptProblem, s::AbstractGradientSolverState, k
+)
+    # remove oldest/last
     pop!(a.gradients)
     M = get_manifold(mp)
-    step, d = a.direction(mp, s, k) #get inner gradient and step
-    a.gradients = vcat([deepcopy(d)], a.gradients)
-    for i in 1:(length(a.gradients) - 1) #transport & shift in place
-        vector_transport_to!(
-            M,
-            a.gradients[i],
-            a.last_iterate,
-            a.gradients[i + 1],
-            get_iterate(s),
-            a.vector_transport_method,
-        )
+    p = get_iterate(s)
+    _, d = a.direction(mp, s, k) #get inner gradient and step
+    for g in a.gradients
+        vector_transport_to!(M, g, a.last_iterate, g, p, a.vector_transport_method)
     end
-    a.gradients[1] = deepcopy(d)
-    copyto!(M, a.last_iterate, get_iterate(s))
-    return step, 1 / length(a.gradients) .* sum(a.gradients)
+    pushfirst!(a.gradients, copy(M, p, d))
+    copyto!(M, a.last_iterate, p)
+    return 1.0, 1 / length(a.gradients) .* sum(a.gradients)
 end
 
-@doc raw"""
-    Nesterov <: DirectionUpdateRule
+"""
+    AverageGradient()
 
-## Fields
+Add an average of gradients to a gradient processor. A set of previous directions (from the
+inner processor) and the last iterate are stored, average is taken after vector transporting
+them to the current iterates tangent space.
+
+# Input
+
+* `M` (optional)
+
+# Keyword arguments
+
+* $(_kw_p_default)
+* `direction=`[`IdentityUpdateRule`](@ref) preprocess the actual gradient before adding momentum
+* `gradients=[zero_vector(M, p) for _ in 1:n]` how to initialise the internal storage
+* `n=10` number of gradient evaluations to take the mean over
+* $(_kw_X_default)`
+* `vector_transport_method=default_vector_transport_method(M, typeof(p)),
+
+!!! info
+    This function generates a [`DirectionUpdateRuleFactory`](@ref) for this
+    for the [`AverageGradientRule`](@ref). If you do not provide the manifold
+    or any other of the keywords that require the manifold for a default,
+    you can provide a manifold later to use that manifold's default's instead.
+"""
+AverageGradient(args...; kwargs...) =
+    DirectionUpdateRuleFactory(AverageGradientRule, args...; kwargs...)
+
+@doc raw"""
+    NesterovRule <: DirectionUpdateRule
+
+Compute a Nesterov inspired direction update rule.
+See [`Nesterov`](@ref) for details
+
+# Fields
+
 * `γ`
 * `μ` the strong convexity coefficient
 * `v` (=``=v_k``, ``v_0=x_0``) an interim point to compute the next gradient evaluation point `y_k`
 * `shrinkage` (`= i -> 0.8`) a function to compute the shrinkage ``β_k`` per iterate.
 
-Assume ``f`` is ``L``-Lipschitz and ``μ``-strongly convex. Given
-
-* a step size ``h_k<\frac{1}{L}`` (from the [`GradientDescentState`](@ref)
-* a `shrinkage` parameter ``β_k``
-* and a current iterate ``x_k``
-* as well as the interim values ``γ_k`` and ``v_k`` from the previous iterate.
-
-This compute a Nesterov type update using the following steps, see [ZhangSra:2018](@cite)
-
-1. Compute the positive root ``α_k∈(0,1)`` of ``α^2 = h_k\bigl((1-α_k)γ_k+α_k μ\bigr)``.
-2. Set ``\bar γ_k+1 = (1-α_k)γ_k + α_kμ``
-3. ``y_k = \operatorname{retr}_{x_k}\Bigl(\frac{α_kγ_k}{γ_k + α_kμ}\operatorname{retr}^{-1}_{x_k}v_k \Bigr)``
-4. ``x_{k+1} = \operatorname{retr}_{y_k}(-h_k \operatorname{grad}f(y_k))``
-5. ``v_{k+1} = `\operatorname{retr}_{y_k}\Bigl(\frac{(1-α_k)γ_k}{\barγ_k}\operatorname{retr}_{y_k}^{-1}(v_k) - \frac{α_k}{\bar γ_{k+1}}\operatorname{grad}f(y_k) \Bigr)``
-6. ``γ_{k+1} = \frac{1}{1+β_k}\bar γ_{k+1}``
-
-Then the direction from ``x_k`` to ``x_k+1`` by ``d = \operatorname{retr}^{-1}_{x_k}x_{k+1}`` is returned.
-
 # Constructor
-    Nesterov(M::AbstractManifold, p::P; γ=0.001, μ=0.9, shrinkage = k -> 0.8;
-        inverse_retraction_method=LogarithmicInverseRetraction())
 
-Initialize the Nesterov acceleration, where `x0` initializes `v`.
 """
-mutable struct Nesterov{P,R<:Real} <: DirectionUpdateRule
+mutable struct NesterovRule{P,R<:Real} <: DirectionUpdateRule
     γ::R
     μ::R
     v::P
     shrinkage::Function
     inverse_retraction_method::AbstractInverseRetractionMethod
 end
-function Nesterov(
-    M::AbstractManifold,
-    p::P;
+function NesterovRule(
+    M::AbstractManifold;
+    p::P=rand(M),
     γ::T=0.001,
     μ::T=0.9,
     shrinkage::Function=i -> 0.8,
     inverse_retraction_method::AbstractInverseRetractionMethod=default_inverse_retraction_method(
-        M, P
+        M, typeof(p)
     ),
 ) where {P,T}
     p_ = _ensure_mutating_variable(p)
-    return Nesterov{typeof(p_),T}(γ, μ, copy(M, p_), shrinkage, inverse_retraction_method)
+    return NesterovRule{typeof(p_),T}(
+        γ, μ, copy(M, p_), shrinkage, inverse_retraction_method
+    )
 end
-function (n::Nesterov)(mp::AbstractManoptProblem, s::AbstractGradientSolverState, k)
+function (n::NesterovRule)(mp::AbstractManoptProblem, s::AbstractGradientSolverState, k)
     M = get_manifold(mp)
     h = get_stepsize(mp, s, k)
     p = get_iterate(s)
@@ -573,6 +584,48 @@ function (n::Nesterov)(mp::AbstractManoptProblem, s::AbstractGradientSolverState
     n.v = retract(M, y, d, s.retraction_method)
     n.γ = 1 / (1 + n.shrinkage(k)) * γbar
     return h, (-1 / h) * inverse_retract(M, p, xn, n.inverse_retraction_method) # outer update
+end
+@doc """
+    Nesterov()
+
+Assume ``f`` is ``L``-Lipschitz and ``μ``-strongly convex. Given
+
+* a step size ``h_k<$(_l[:frac]("1","L"))`` (from the [`GradientDescentState`](@ref)
+* a `shrinkage` parameter ``β_k``
+* and a current iterate ``p_k``
+* as well as the interim values ``γ_k`` and ``v_k`` from the previous iterate.
+
+This compute a Nesterov type update using the following steps, see [ZhangSra:2018](@cite)
+
+1. Compute the positive root ``α_k∈(0,1)`` of ``α^2 = h_k$(_l[:bigl])((1-α_k)γ_k+α_k μ$(_l[:bigr]))``.
+2. Set ``$(_l[:bar]("γ"))_k+1 = (1-α_k)γ_k + α_kμ``
+3. ``y_k = $(_l_retr)_{p_k}\\Bigl(\\frac{α_kγ_k}{γ_k + α_kμ}$(_l_retr)^{-1}_{p_k}v_k \\Bigr)``
+4. ``x_{k+1} = $(_l_retr)_{y_k}(-h_k $(_l_grad)f(y_k))``
+5. ``v_{k+1} = $(_l_retr)_{y_k}\\Bigl(\\frac{(1-α_k)γ_k}{$(_l[:bar]("γ"))_k}$(_l_retr)_{y_k}^{-1}(v_k) - \\frac{α_k}{$(_l[:bar]("γ"))_{k+1}}$(_l_grad)f(y_k) \\Bigr)``
+6. ``γ_{k+1} = \\frac{1}{1+β_k}$(_l[:bar]("γ"))_{k+1}``
+
+Then the direction from ``p_k`` to ``p_k+1`` by ``d = $(_l_retr)^{-1}_{p_k}p_{k+1}`` is returned.
+
+# Input
+
+* `M` (optional)
+
+# Keyword arguments
+
+* $(_kw_p_default)
+* `γ=0.001``
+* `μ=0.9``
+* `shrinkage = k -> 0.8`
+* $(_kw_inverse_retraction_method_default): $(_kw_inverse_retraction_method)
+
+!!! info
+    This function generates a [`DirectionUpdateRuleFactory`](@ref) for this
+    for the [`NesterovRule`](@ref). If you do not provide the manifold
+    or any other of the keywords that require the manifold for a default,
+    you can provide a manifold later to use that manifold's default's instead.
+"""
+function Nesterov(args...; kwargs...)
+    return DirectionUpdateRuleFactory(NesterovRule, args...; kwargs...)
 end
 
 @doc raw"""